crmproject/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.2
django-auditlog==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
